@@ -26,14 +26,17 @@ android {
             storePassword 'madeinhk'
         }
     }
-    compileSdkVersion 23
-    buildToolsVersion "23.0.1"
+    compileSdkVersion 24
+    buildToolsVersion "24.0.2"
     defaultConfig {
         applicationId "com.madeinhk.english_chinesedictionary"
         minSdkVersion 16
-        targetSdkVersion 23
-        versionCode 26
-        versionName "0.6.6"
+        targetSdkVersion 24
+        versionCode 27
+        versionName "0.6.7"
+        jackOptions {
+            enabled true
+        }
     }
     buildTypes {
         release {
@@ -47,8 +50,8 @@ android {
         }
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_7
-        targetCompatibility JavaVersion.VERSION_1_7
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
     }
 
     lintOptions {
@@ -58,15 +61,16 @@ android {
 
 dependencies {
     compile fileTree(include: ['*.jar'], dir: 'libs')
-    compile 'com.android.support:preference-v14:23+'
-    compile 'com.android.support:appcompat-v7:23+'
+    compile 'com.android.support:preference-v14:24+'
+    compile 'com.android.support:appcompat-v7:24+'
     compile 'de.greenrobot:eventbus:2.4.0'
-    compile 'com.android.support:recyclerview-v7:23+'
-    compile 'com.android.support:cardview-v7:23+'
+    compile 'com.android.support:recyclerview-v7:24+'
+    compile 'com.android.support:cardview-v7:24+'
     compile 'com.google.android.gms:play-services-analytics:+'
     compile project(':sqlite-assest-helper')
-    compile 'com.android.support:recyclerview-v7:23+'
-    compile 'com.android.support:design:23+'
+    compile 'com.android.support:recyclerview-v7:24+'
+    compile 'com.android.support:design:24+'
+    compile 'com.android.support:support-compat:24+'
     compile 'org.jsoup:jsoup:+'
     compile('com.crashlytics.sdk.android:crashlytics:2.3.2@aar') {
         transitive = true;
