@@ -1,6 +1,7 @@
 package com.madeinhk.english_chinesedictionary.service;
 
-import android.annotation.SuppressLint;
+import android.app.Notification;
+import android.app.NotificationManager;
 import android.app.Service;
 import android.content.ClipData;
 import android.content.ClipDescription;
@@ -9,11 +10,15 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Build;
+import android.os.Bundle;
 import android.os.IBinder;
+import android.support.v4.app.RemoteInput;
+import android.support.v4.os.BuildCompat;
 import android.support.v7.preference.PreferenceManager;
 import android.text.TextUtils;
 import android.util.Log;
 
+import com.madeinhk.utils.QuickLookupNotificationHelper;
 import com.madeinhk.english_chinesedictionary.SettingFragment;
 import com.madeinhk.model.ECDictionary;
 import com.madeinhk.model.Word;
@@ -22,16 +27,31 @@ import com.madeinhk.utils.StringUtils;
 
 import org.jsoup.Jsoup;
 
-public class ClipboardService extends Service {
+public class ECDictionaryService extends Service {
+    private static final String TAG = "ECDictionaryService";
     private boolean mRegistered = false;
 
     private String mLastWord;
     private long mTimestamp;
 
-    public ClipboardService() {
+    private static int NOTIFICATION_ID = 1;
+
+    public static final String KEY_TEXT_REPLY = "key_text_reply";
+    public static final String ACTION_QUICK_LOOKUP = "quick_lookup";
+    public static final String KEY_IS_FOREGROUND = "key_is_foreground";
+    public static final String ACTION_CHANGE_FOREGROUND = "change_foreground";
+
+    private QuickLookupNotificationHelper mQuickLookHelper;
+
+    public ECDictionaryService() {
     }
 
     public void onCreate() {
+        Log.d(TAG, "onCreate: " + mRegistered);
+        if (BuildCompat.isAtLeastN()) {
+            mQuickLookHelper = new QuickLookupNotificationHelper(this);
+            startForeground();
+        }
         synchronized (mClipListener) {
             if (!mRegistered) {
                 ClipboardManager clipboardManager =
@@ -42,20 +62,44 @@ public class ClipboardService extends Service {
         }
     }
 
+    private void startForeground() {
+        Notification notification = mQuickLookHelper.buildInitialNotification();
+        startForeground(NOTIFICATION_ID, notification);
+    }
+
     public static void start(Context context) {
-        context.startService(new Intent(context.getApplicationContext(), ClipboardService.class));
+        context.startService(
+                new Intent(context.getApplicationContext(), ECDictionaryService.class));
     }
 
     @Override
     public int onStartCommand(Intent intent, int flags, int startId) {
+        if (BuildCompat.isAtLeastN() && ACTION_QUICK_LOOKUP.equals(intent.getAction())) {
+            String text = getMessageText(intent);
+            if (text != null) {
+                text = text.trim();
+            }
+            Word word = lookupWord(text);
+            Notification notification = mQuickLookHelper.buildNotificationForResult(word);
+            NotificationManager notificationManager =
+                    (NotificationManager) getSystemService(Context.NOTIFICATION_SERVICE);
+            notificationManager.notify(NOTIFICATION_ID, notification);
+        } else if (ACTION_CHANGE_FOREGROUND.equals(intent.getAction())) {
+            boolean isForeground = intent.getBooleanExtra(KEY_IS_FOREGROUND, true);
+            if (isForeground) {
+                startForeground();
+            } else {
+                stopForeground(true);
+                stopSelf();
+            }
 
+        }
         return START_STICKY;
     }
 
     @Override
     public IBinder onBind(Intent intent) {
-        // TODO: Return the communication channel to the service.
-        throw new UnsupportedOperationException("Not yet implemented");
+        throw null;
     }
 
     ClipboardManager.OnPrimaryClipChangedListener mClipListener =
@@ -77,24 +121,16 @@ public class ClipboardService extends Service {
                     if (text != null) {
                         text = text.trim();
                     }
-
-                    if (!TextUtils.isEmpty(text) && !isDuplicated(text)) {
-                        ECDictionary dictionary = new ECDictionary(ClipboardService.this);
-                        String str = text.toLowerCase();
-                        mLastWord = str;
-                        mTimestamp = System.currentTimeMillis();
-                        Word word = dictionary.lookup(str);
-                        if (word == null && StringUtils.isEnglishWord(str)) {
-                            // Try to have stemming
-                            Stemmer stemmer = new Stemmer();
-                            stemmer.add(str.toCharArray(), str.length());
-                            stemmer.stem();
-                            word = dictionary.lookup(stemmer.toString());
-                        }
-                        if (word != null) {
-                            DictionaryHeadService.show(ClipboardService.this, word);
-                        }
+                    if (TextUtils.isEmpty(text) || isDuplicated(text)) {
+                        return;
                     }
+                    mLastWord = text;
+                    mTimestamp = System.currentTimeMillis();
+                    Word word = lookupWord(text);
+                    if (word != null) {
+                        DictionaryHeadService.show(ECDictionaryService.this, word);
+                    }
+
                 }
 
                 // To tackle with the funny behaviour of clipboard listener
@@ -103,7 +139,6 @@ public class ClipboardService extends Service {
                 }
 
                 // It seems that the clipboard api is quite buggy, defensive coding here
-                @SuppressLint("NewApi")
                 private String extractTextFromClipData(ClipData clipData) {
                     if (clipData == null) {
                         return null;
@@ -143,8 +178,47 @@ public class ClipboardService extends Service {
         }
     }
 
+    private Word lookupWord(String text) {
+        if (TextUtils.isEmpty(text)) {
+            return null;
+        }
+        ECDictionary dictionary = new ECDictionary(ECDictionaryService.this);
+        String str = text.toLowerCase();
+        Word word = dictionary.lookup(str);
+        if (word != null) {
+            return word;
+        }
+        if (StringUtils.isEnglishWord(str)) {
+            // Try to have stemming
+            Stemmer stemmer = new Stemmer();
+            stemmer.add(str.toCharArray(), str.length());
+            stemmer.stem();
+            return dictionary.lookup(stemmer.toString());
+        }
+        return null;
+    }
+
     private boolean isFeatureEnabled() {
         SharedPreferences preferences = PreferenceManager.getDefaultSharedPreferences(this);
         return preferences.getBoolean(SettingFragment.KEY_COPY_TO_LOOKUP, true);
+    }
+
+    private String getMessageText(Intent intent) {
+        Bundle remoteInput = RemoteInput.getResultsFromIntent(intent);
+        if (remoteInput != null) {
+            return remoteInput.getCharSequence(KEY_TEXT_REPLY).toString();
+        }
+        return null;
+    }
+
+    public static Intent getQuickLookupIntent(Context context) {
+        return new Intent(context, ECDictionaryService.class).setAction(ACTION_QUICK_LOOKUP);
+    }
+
+    public static Intent getChangeForegroundIntent(Context context, boolean isForeground) {
+        Intent intent = new Intent(context, ECDictionaryService.class).
+                setAction(ACTION_CHANGE_FOREGROUND);
+        intent.putExtra(KEY_IS_FOREGROUND, isForeground);
+        return intent;
     }
 }
